refactor(state): See #92 - Setting up the state before testing handlersto avoid having them tied to component
