+++ src/state/index.ts
@@ -5,7 +5,8 @@
 import { useImmer } from "use-immer";
 import useWebsocket from "./websocket";
 import useUser from "./user";
-import { genId, getUrlParams, isSSR } from "../utils";
+import { decodeMessage, encodeMessage, genId, getUrlParams, isSSR, verifySignatureFromPeerId } from "../utils";
+import { API } from "../lib/api";
 
 const MYNE_CHAT_GIT_HASH = process.env.NEXT_PUBLIC_MYNE_CHAT_GIT_HASH
 const MYNE_CHAT_VERSION = require('../../package.json').version
@@ -14,6 +15,8 @@ const MYNE_CHAT_ENVIRONMENT = process.env.NEXT_PUBLIC_MYNE_CHAT_ENVIRONMENT
 export type { ConnectionStatus } from "./websocket";
 
 export type VerifiedStatus = "UNVERIFIED" | "VERIFIED" | "FAILED_VERIFICATION"
+
+export type UpdateMessageHandlerInterface = (counterparty: string, messageId: string, status: Message["status"], error?: string | undefined) => void
 
 export type Message = {
   id: string;
@@ -163,21 +166,78 @@ const useAppState = () => {
     });
   };
 
+  const handleAddNewConversation = (callback: () => void) => (counterparty: string) => {
+    addNewConversation(counterparty);
+    callback();
+  };
+
+  const handleSendMessage = async (destination: string, message: string) => {
+    const { selection, settings, verified } = state;
+    const { myPeerId } = user?.state;
+    const { socketRef } = websocket;
+    if (!myPeerId || !selection || !socketRef.current) return;
+
+    const headers = user.getReqHeaders(true)
+    const api = API(settings.httpEndpoint, headers)
+    const signature = verified && await api.signRequest(message)
+      .catch((err: any) => console.error('ERROR Failed to obtain signature', err));
+    const encodedMessage = encodeMessage(myPeerId, message, signature);
+    const id = addSentMessage(myPeerId, destination, message);
+
+    await api.sendMessage(selection, encodedMessage, destination, id, updateMessage)
+      .catch((err: any) => console.error('ERROR Failed to send message', err));
+  };
+
+  const handleReceivedMessage = async (ev: MessageEvent<string>) => {
+    try {
+      // we are only interested in messages, not all the other events coming in on the socket
+      const data = JSON.parse(ev.data);
+      if (data.type == "message") {
+        const { tag, from, message, signature } = decodeMessage(data.msg);
+
+        const verifiedStatus : VerifiedStatus = signature ? 
+          (await verifySignatureFromPeerId(from, message, signature) ? 
+            "VERIFIED" :
+            "FAILED_VERIFICATION"
+          ) :
+          "UNVERIFIED";
+
+        // we are only interested in myne messages
+        if (tag == "myne") {
+          addReceivedMessage(from, message, verifiedStatus);
+        }
+      }
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
+  const loadDevHelperConversation = () => {
+    console.log("⚙️  Developer Mode enabled.", process.env.NODE_ENV)
+    const dev = '⚙️  Dev'
+    addNewConversation(dev)
+    // setTimeout ensures the event loop takes these state updates in order.
+    setTimeout(() => addReceivedMessage(dev, 'Welcome to the developer mode.'), 0)
+    setTimeout(() => addReceivedMessage(dev, 'This conversation is only available during development.'), 0)
+    setTimeout(() => addReceivedMessage(dev, 'This is how a verified message looks like.', 'VERIFIED'), 0)
+    setTimeout(() => addReceivedMessage(dev, 'This is how an unverified message looks like.', 'UNVERIFIED'), 0)
+    setTimeout(() => addReceivedMessage(dev, 'This is how a failed verification message looks like.', 'FAILED_VERIFICATION'), 0)
+  }
+
   return {
     state: {
       ...state,
       ...websocket.state,
       ...user.state,
     },
-    getReqHeaders: user.getReqHeaders,
     socketRef: websocket.socketRef,
     updateSettings,
     setSelection,
     setVerified,
-    addSentMessage,
-    addReceivedMessage,
-    updateMessage,
-    addNewConversation,
+    handleAddNewConversation,
+    handleSendMessage,
+    handleReceivedMessage,
+    loadDevHelperConversation,
     hash: MYNE_CHAT_GIT_HASH,
     version: MYNE_CHAT_VERSION,
     environment: MYNE_CHAT_ENVIRONMENT
